move clinvar to reference data
